Moves plotting functionality into ca_utils file so that superpixel_analysis_ring only contains the core demixing pipeline now
